Separate logger and loggerprovider mod
